Broke classes out into their own files. Fully tested the utilities methods.
