+++ scripts/run_tests.sh
@@ -1,0 +1,1 @@
+python3 -m unittest -v src/tests/*_test.py
